Move submission routes to their own module
