+++ initialization/routes.py
@@ -7,6 +7,7 @@ import pages.static as static_pages
 import pages.strains as strain_pages
 import pages.studies as study_pages
 import pages.upload as upload_pages
+import pages.submission as submission_pages
 
 
 def init_routes(app):
@@ -25,9 +26,21 @@ def init_routes(app):
     app.add_url_rule("/upload/4", view_func=upload_pages.upload_step4_page, methods=["GET", "POST"])
     app.add_url_rule("/upload/5", view_func=upload_pages.upload_step5_page, methods=["GET", "POST"])
 
-    app.add_url_rule("/upload/new_submission",         view_func=upload_pages.new_submission_action,    methods=["POST"])
-    app.add_url_rule("/upload/edit_submission/<id>",   view_func=upload_pages.edit_submission_action,   methods=["POST"])
-    app.add_url_rule("/upload/delete_submission/<id>", view_func=upload_pages.delete_submission_action, methods=["POST"])
+    app.add_url_rule(
+        "/upload/new_submission",
+        view_func=submission_pages.new_submission_action,
+        methods=["POST"],
+    )
+    app.add_url_rule(
+        "/upload/edit_submission/<id>",
+        view_func=submission_pages.edit_submission_action,
+        methods=["POST"],
+    )
+    app.add_url_rule(
+        "/upload/delete_submission/<id>",
+        view_func=submission_pages.delete_submission_action,
+        methods=["POST"],
+    )
 
     app.add_url_rule("/upload/study_template.xlsx", view_func=upload_pages.upload_study_template_xlsx)
     app.add_url_rule(
